+++ CHANGELOG.md
@@ -17,6 +17,7 @@
 ### Changed
 - CiA 301 PDO related registers are not saved to the configuration file. 
 - Remove an unnecessary argument for a servo status listener callback.
+- DictionaryV2 raises an exception if the interface does not match.
 
 ## [7.3.5] - 2025-08-23